Update sko1-skj.json
the correct options to question:
What phenomena will occur when we convert the concentrator (hub) into a store-and-forward switch in the Ethernet network?
+ The delay of transmission between nodes will increase
- The maximum possible distance between nodes will be reduced
+ Thanks to the reduction of the number of collisions, the number of lost frames will decrease
+ Increases bisective bandwidth of the network (sumacary bandwidth, when half of the nodes broadcasts, half receives)
- Some protocols, for example. ARP will stop working properly


should be first, 3rd and 4th options, however, in past it  there was selected 2nd and 3rd in where 2nd is definitely wrong. 


Justification:
1st correct because:
In a store-and-forward switch, each Ethernet frame is received in its entirety, checked for errors, and then forwarded to the appropriate port. This process adds a small amount of delay compared to a hub, which simply repeats the electrical signal to all ports without examining the frame. The delay is due to the time needed for the switch to process and forward the frame.

2nd wrong becasue:
The maximum cable lengths are generally determined by Ethernet standards and are not significantly impacted by the type of device used (hub or switch).

3rd correct.

4th also correct because:
Switches can increase the bisectional (or aggregate) bandwidth of a network. Since each switch port is a separate collision domain, multiple pairs of devices can communicate simultaneously without interfering with each other.